requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
transformers>=4.30.0

# Evaluation metrics
sacrebleu>=2.3.0
nltk>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
